ON-17161: include file handles in shrub client's "mappings"
As far as I can figure out, there's no way to map hugetlb pages into
user memory, except by mapping the file they came from. So we'll need
access to at least the "buffers" file in order to map a shared onload
stack. We might as well share all the file handles, and map all the
shared memory in the same manner.

+++ src/lib/ciul/shrub_socket.c
@@ -124,7 +124,7 @@ int ef_shrub_socket_recv(uintptr_t socket, void* data, size_t bytes)
 }
 
 int ef_shrub_socket_recv_metrics(struct ef_shrub_shared_metrics* metrics,
-                                 uintptr_t* shared_files,
+                                 uint64_t* shared_files,
                                  uintptr_t socket)
 {
   int rc, i;
